+++ CHANGELOG.md
@@ -9,10 +9,11 @@ Repository: **<git+https://github.com/vladmandic/human.git>**
 
 ## Changelog
 
+### **HEAD -> main** 2021/05/11 mandic00@live.com
+
+- fix node build and update model signatures
+
 ### **1.8.4** 2021/05/11 mandic00@live.com
-
-
-### **origin/main** 2021/05/05 mandic00@live.com
 
 
 ### **1.8.3** 2021/05/05 mandic00@live.com