+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @vladmandic/human  
 
-Version: **1.8.4**  
+Version: **1.8.5**  
 Description: **Human: AI-powered 3D Face Detection & Rotation Tracking, Face Description & Recognition, Body Pose Tracking, 3D Hand & Finger Tracking, Iris Analysis, Age & Gender & Emotion Prediction, Gesture Recognition**  
 
 Author: **Vladimir Mandic <mandic00@live.com>**  
@@ -9,11 +9,12 @@ Repository: **<git+https://github.com/vladmandic/human.git>**
 
 ## Changelog
 
-### **HEAD -> main** 2021/05/17 mandic00@live.com
+### **1.8.5** 2021/05/18 mandic00@live.com
 
 
-### **origin/main** 2021/05/16 mandic00@live.com
+### **origin/main** 2021/05/17 mandic00@live.com
 
+- add node-video sample
 - add node-webcam demo
 - fix node build and update model signatures